win32_*.c: avoid some format string warnings
These aren't real issues as all the types in question are 32-bit ints,
but fixing these will prevent warnings later if we're ever able to
enable __attribute__((format)) on Windows (hopefully due to
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=64862 being fixed).

+++ src/win32_apply.c
@@ -1480,7 +1480,7 @@ retry:
 	if (unlikely(!NT_SUCCESS(status))) {
 		winnt_error(status, L"Can't open \"%ls\" for deletion "
 			    "(perms=%x, flags=%x)",
-			    current_path(ctx), perms, flags);
+			    current_path(ctx), (u32)perms, (u32)flags);
 		return WIMLIB_ERR_OPEN;
 	}
